Correction in matrix.h

+++ src/1.cpp
@@ -3,7 +3,7 @@
 
 int main()
 {
-    matrix<int> matrx1, matrx2;
+    matrix<float> matrx1, matrx2;
     std::cout << "Enter the matrix 1(First number of rows and columns and then the elements:\n";
     std::cin >> matrx1;
     std::cout << "Enter the matrix 1(First number of rows and columns and then the elements:\n";
